search fix, routing update
